[macos] Make tmp copy of file at `previousUploadURL` before passing into `NSItemProvider(contentsOf: ...)`
Previously, the file at `previousUploadURL` would get deleted after the call to `NSItemProvider(contentsOf: ...)`. This meant that the first call to `.onDrag` would succeed, but subsequent attempts would fail as the file no longer existed.

With this approach we create a temporary copy of the file at `previousUploadURL` every single time `.onDrag` (and subsequently `NSItemProvider(contentsOf: ...)`) are called... ensuring that we can "drag" a file multiple times.

This solution isn't great as it requires writing a new file to disk every single time we want to "drag" a file. In the future we may instead be able to allocate an `NSImage` and use `NSItemProvider(object: <#T##NSItemProviderWriting#>)` instead. However, `NSImage` doesn't implement `NSItemProviderWriting` until macOS 13 Ventura. So this current solution seems to work fine for now? Things remain interactive for now despite having to write a file to disk just to immediately pass to `NSItemProvider` and then be deleted. This probably **would** be an issue with larger files though?