[macos] Create copy of image to pass to `UNNotificationAttachment`
When we pass the local image path to `UNNotificationAttachment` it appears to copy the image from `NSTemporaryDirectory()` to some sort of stable attachment store. It looks like this is to ensure that the image exists and can be displayed when the notification is visible. At least that's what I gathered reading eg:

> Note that iOS automatically moves the file into the attachment data store, so this file will be deleted from the temporary directory right away. You don't need to worry about clean up. (https://stackoverflow.com/questions/39103095/unnotificationattachment-with-uiimage-or-remote-url#comment106978607_39103096)

However, we depend on the image being available locally so we can display it in the app. So, just like we do w/ `onDrag`, we create a copy of the file and pass that to `UNNotificationAttachment` instead.

We probably should store uploads in Documents directory instead of `NSTemporaryDirectory` since w/ `NSTemporaryDirectory` we have no guarantees or concrete idea of when files will be "cleaned up."