[macos] Replace `@Observable` macro with `@Model` for `BlobEntry`
As part of trying out SwiftData for persistence.

Made some additional changes to `AppDelegate` to accomodate this change by creating a `modelContainer`, this prevents fatal error when a `BlobEntry` object is instantiated.

There is no actual persistence via SwiftData at the moment, but the "model" is ready. We'll need to configure the container/context next.